API: allow user to delete its characters

+++ src/Tavernkeep.Application/UseCases/Characters/Commands/DeleteCharacter/DeleteCharacterCommandHandler.cs
@@ -1,15 +1,22 @@
 ﻿using MediatR;
+using Tavernkeep.Core.Contracts.Enums;
 using Tavernkeep.Core.Exceptions;
 using Tavernkeep.Core.Repositories;
 
 namespace Tavernkeep.Application.Actions.Characters.Commands.DeleteCharacter
 {
-    public class DeleteCharacterCommandHandler(ICharacterRepository characterRepository) : IRequestHandler<DeleteCharacterCommand>
+    public class DeleteCharacterCommandHandler(IUserRepository userRepository, ICharacterRepository characterRepository) : IRequestHandler<DeleteCharacterCommand>
     {
         public async Task Handle(DeleteCharacterCommand request, CancellationToken cancellationToken)
         {
+            var initiator = await userRepository.FindAsync(request.InitiatorId)
+                ?? throw new BusinessLogicException("User with specified ID doesn't exist.");
+
             var character = await characterRepository.FindAsync(request.CharacterId)
                 ?? throw new BusinessLogicException("Character not found");
+
+            if (character.Owner.Id != request.InitiatorId && initiator.Role != UserRole.Master)
+                throw new InsufficientPermissionException("You do not have the necessary permissions to perform this operation.");
 
             characterRepository.Remove(character);
             await characterRepository.CommitAsync(cancellationToken);
